requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-debug-toolbar==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
flake8==6.0.0
flake8-quotes==3.3.2
flake8-variables-names==0.0.6
idna==3.4
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
mccabe==0.7.0
pep8-naming==0.13.3
Pillow==10.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
rpds-py==0.9.2
sqlparse==0.4.4
stripe==5.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
